+++ packages/forest_admin_agent/lib/forest_admin_agent/utils/schema/schema_emitter.rb
@@ -6,7 +6,7 @@ module ForestAdminAgent
     module Schema
       class SchemaEmitter
         LIANA_NAME = "agent-ruby"
-        LIANA_VERSION = "1.39.4"
+        LIANA_VERSION = "1.40.1"
 
         def self.generate(datasource)
           datasource.collections