+++ packages/forest_admin_agent/spec/lib/forest_admin_agent/routes/resources/related/associate_related_spec.rb
@@ -81,6 +81,15 @@ module ForestAdminAgent
                     origin_key_target: 'id',
                     origin_type_field: 'addressable_type',
                     origin_type_value: 'user'
+                  ),
+                  'locked_addresses' => Relations::ManyToManySchema.new(
+                    foreign_key: 'address_id',
+                    foreign_collection: 'address',
+                    foreign_key_target: 'id',
+                    through_collection: 'address_user',
+                    origin_key: 'user_id',
+                    origin_key_target: 'id',
+                    is_read_only: true
                   )
                 }
               }
@@ -189,6 +198,19 @@ module ForestAdminAgent
               expect(action).to eq(:edit)
               expect(collection.name).to eq('address')
             end
+          end
+
+          it 'refuses to associate a read-only relation (any type, not just OneToOne, can be ' \
+             'marked is_read_only now that it lives on the shared RelationSchema)' do
+            args[:params]['relation_name'] = 'locked_addresses'
+            args[:params]['data'] = [{ 'id' => 1 }]
+            args[:params]['id'] = 1
+            allow(@datasource.get_collection('address_user')).to receive(:create)
+
+            expect { associate.handle_request(args) }
+              .to raise_error(ForestAdminDatasourceToolkit::Exceptions::ValidationError,
+                              'Field locked_addresses is not editable')
+            expect(@datasource.get_collection('address_user')).not_to have_received(:create)
           end
 
           context 'when call on one to many relation' do