+++ packages/forest_admin_datasource_active_record/spec/dummy/app/models/parent.rb
@@ -1,0 +1,4 @@
+class Parent < ApplicationRecord
+  has_many :kids, as: :owner
+  has_many :details, through: :kids, source: :detail
+end