+++ packages/forest_admin_datasource_active_record/spec/dummy/app/models/solo.rb
@@ -1,0 +1,4 @@
+class Solo < ApplicationRecord
+  has_one :kid, as: :owner
+  has_one :detail, through: :kid
+end