+++ packages/forest_admin_datasource_active_record/spec/dummy/app/models/tag.rb
@@ -1,0 +1,3 @@
+class Tag < ApplicationRecord
+  belongs_to :slot, foreign_key: :custom_slot_id, inverse_of: :tag
+end