+++ packages/forest_admin_datasource_toolkit/lib/forest_admin_datasource_toolkit/schema/relations/one_to_many_schema.rb
@@ -4,8 +4,8 @@ module ForestAdminDatasourceToolkit
       class OneToManySchema < RelationSchema
         attr_accessor :origin_key, :origin_key_target
 
-        def initialize(origin_key:, origin_key_target:, foreign_collection:)
-          super(foreign_collection, 'OneToMany')
+        def initialize(origin_key:, origin_key_target:, foreign_collection:, is_read_only: false)
+          super(foreign_collection, 'OneToMany', is_read_only: is_read_only)
           @origin_key = origin_key
           @origin_key_target = origin_key_target
         end